// web-application/src/types-openapi/api/models/UpdateMembershipRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateMembershipRequest
 */
export interface UpdateMembershipRequest {
    /**
     * 
     * @type {number}
     * @memberof UpdateMembershipRequest
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateMembershipRequest
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateMembershipRequest
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateMembershipRequest
     */
    discountPercent?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateMembershipRequest
     */
    description?: string | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateMembershipRequest
     */
    durationDays?: number;
    /**
     * 
     * @type {string}
     * @memberof UpdateMembershipRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the UpdateMembershipRequest interface.
 */
export function instanceOfUpdateMembershipRequest(value: object): value is UpdateMembershipRequest {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function UpdateMembershipRequestFromJSON(json: any): UpdateMembershipRequest {
    return UpdateMembershipRequestFromJSONTyped(json, false);
}

export function UpdateMembershipRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateMembershipRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'price': json['price'] == null ? undefined : json['price'],
        'discountPercent': json['discountPercent'] == null ? undefined : json['discountPercent'],
        'description': json['description'] == null ? undefined : json['description'],
        'durationDays': json['durationDays'] == null ? undefined : json['durationDays'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function UpdateMembershipRequestToJSON(json: any): UpdateMembershipRequest {
    return UpdateMembershipRequestToJSONTyped(json, false);
}

export function UpdateMembershipRequestToJSONTyped(value?: UpdateMembershipRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'price': value['price'],
        'discountPercent': value['discountPercent'],
        'description': value['description'],
        'durationDays': value['durationDays'],
        'status': value['status'],
    };
}